// src/lib/models/TerrainResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TerrainResponse = {
    nom: string;
    description?: (string | null);
    localite_id: string;
    id: string;
    user_id: string;
    nombre_parcelles?: (number | null);
    created_at: string;
    updated_at: string;
    superficie?: number;
};